[mod/#76] jobId를 Long에서 Int로

+++ app/src/main/java/com/haphap/app/presentation/register/RegisterContract.kt
@@ -46,7 +46,7 @@ sealed interface RegisterContract {
 
     sealed interface RegisterSideEffect {
         data object Home : RegisterSideEffect
-        data class JobDetail(val jobId: Long) : RegisterSideEffect
+        data class JobDetail(val jobId: Int) : RegisterSideEffect
     }
 
     sealed class SideEffect {
